refractor: moving all-feature to base
